+++ src/pdf/filt_predict.h
@@ -21,7 +21,7 @@ pdf_filt_predict_decode(uint8_t *in,
                         uint8_t colors,
                         uint8_t bpc,
                         uint16_t columns,
-                        uint16_t rows );
+                        size_t rows );
 
 // Encode None, TIFF or PNG predictors
 DLLEXPORT void
@@ -31,6 +31,6 @@ pdf_filt_predict_encode(uint8_t *in,
                         uint8_t colors,
                         uint8_t bpc,
                         uint16_t columns,
-                        uint16_t rows );
+                        size_t rows );
 
 #endif
